Removed code for Go < 1.20
