+++ .github/workflows/build.yml
@@ -13,7 +13,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        go: ["1.22", "1.23", "1.24"]
+        go: ["1.21", "1.22", "1.23", "1.24"]
     steps:
       - name: Checkout
         uses: actions/checkout@v4
